Switch inaccurate BuiltInStructureSets to MojMap's StructureSetKeys (#160)
I think I did this pr correctly but lmk if it is wrong. Basically, this class holds only registry keys for structure sets. Not the actual structure sets (the class StructureSets holds the structure sets). This makes the original name no good. Solution, switch to MojMap name that is more clear that it is holding keys.

(change will need to be propagated to snapshot versions too. Do I need to make a pr for snapshot too?)